+++ src/msg/json/msg_json.cpp
@@ -645,7 +645,7 @@ namespace msg::json
     {
         msg.reserve(1024);
         msg += "[";
-        for (int i = 0; i < instances.size(); i++)
+        for (size_t i = 0; i < instances.size(); i++)
         {
             const hp::instance_info &instance = instances[i];
             msg += "{\"";
